fix: 仅在发送动态视频时删除文件以避免没文件删

+++ module/business/bilibili/push.js
@@ -208,7 +208,7 @@ export default class Bilibilipush extends Base {
             } catch (error) {
               logger.error(error)
             } finally {
-              await this.removeFile(video?.filepath)
+              if(send && Config.bilibili.senddynamicvideo) await this.removeFile(video?.filepath)
             }
           }
